+++ public/main.js
@@ -2112,7 +2112,7 @@ var ViewPatientsComponent = /** @class */ (function () {
     ViewPatientsComponent.prototype.deletePatient = function (id) {
         var _this = this;
         var patients = this.patients;
-        if (window.confirm("Are you sure you want to Delete this patient?")) {
+        if (window.confirm("Are you sure you want to Delete this patient and all of its associated Records?")) {
             this.patientService.deletePatient(id)
                 .subscribe(function (data) {
                 if (data.n == 1) {
